refactor: add dependencies for updated GooglePubSubClient interface from common-jvm
