+++ MODULE.bazel
@@ -40,12 +40,14 @@ bazel_dep(
     version = "0.97.0",
     repo_name = "wfa_common_jvm",
 )
-# DO_NOT_SUBMIT (world-federation-of-advertisers/common-jvm#295)
+
+# DO_NOT_SUBMIT (world-federation-of-advertisers/common-jvm#286)
 archive_override(
     module_name = "common-jvm",
-    strip_prefix = "common-jvm-d23dbac4990e80a80260910e13692054c22b8669",
-    urls = "https://github.com/world-federation-of-advertisers/common-jvm/archive/d23dbac4990e80a80260910e13692054c22b8669.tar.gz",
+    strip_prefix = "common-jvm-ec0231b2c9b932379232ade2a8553eba6ba92302",
+    urls = "https://github.com/world-federation-of-advertisers/common-jvm/archive/ec0231b2c9b932379232ade2a8553eba6ba92302.tar.gz",
 )
+
 bazel_dep(
     name = "common-cpp",
     version = "0.13.0",
