--- imports/java/com/rabbitmq/client/BUILD.bazel
@@ -1,7 +1,0 @@
-# TODO(@marcopremier): Remove this once common-jvm PR #286 is merged
-package(default_visibility = ["//visibility:public"])
-
-alias(
-    name = "client",
-    actual = "@maven//:com_rabbitmq_amqp_client",
-)
