+++ src/main/kotlin/org/wfanet/measurement/securecomputation/controlplane/v1alpha/PublisherProvider.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2024 The Cross-Media Measurement Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import com.google.cloud.pubsub.v1.Publisher
+import com.google.pubsub.v1.TopicName
+
+internal fun createProductionPublisher(projectId: String, topicId: String): Publisher {
+  val topicName = TopicName.of(projectId, topicId)
+  return Publisher.newBuilder(topicName).build()
+}
